+++ src/main/java/com/jajaja/domain/cart/controller/CartController.java
@@ -5,7 +5,7 @@ import com.jajaja.domain.cart.dto.CartResponseDto;
 import com.jajaja.domain.cart.service.CartCommandService;
 import com.jajaja.domain.cart.service.CartQueryService;
 import com.jajaja.global.apiPayload.ApiResponse;
-import com.jajaja.global.config.security.annotation.Auth;
+import com.jajaja.global.security.annotation.Auth;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
