+++ src/main/java/com/jajaja/domain/coupon/controller/CouponController.java
@@ -5,7 +5,7 @@ import com.jajaja.domain.coupon.dto.PagingCouponListResponseDto;
 import com.jajaja.domain.coupon.service.CouponCommandService;
 import com.jajaja.domain.coupon.service.CouponQueryService;
 import com.jajaja.global.apiPayload.ApiResponse;
-import com.jajaja.global.config.security.annotation.Auth;
+import com.jajaja.global.security.annotation.Auth;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
