+++ src/main/java/com/jajaja/domain/member/controller/MemberBusinessCategoryController.java
@@ -3,7 +3,7 @@ package com.jajaja.domain.member.controller;
 import com.jajaja.domain.member.dto.MemberBusinessCategoryRequestDto;
 import com.jajaja.domain.member.service.MemberBusinessCategoryCommandService;
 import com.jajaja.global.apiPayload.ApiResponse;
-import com.jajaja.global.config.security.annotation.Auth;
+import com.jajaja.global.security.annotation.Auth;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
