+++ src/main/java/com/jajaja/domain/member/controller/MemberController.java
@@ -5,7 +5,7 @@ import com.jajaja.domain.member.dto.response.MemberInfoResponseDto;
 import com.jajaja.domain.member.service.MemberCommandService;
 import com.jajaja.domain.member.service.MemberQueryService;
 import com.jajaja.global.apiPayload.ApiResponse;
-import com.jajaja.global.config.security.annotation.Auth;
+import com.jajaja.global.security.annotation.Auth;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
