+++ src/main/java/com/jajaja/domain/team/controller/TeamController.java
@@ -5,7 +5,7 @@ import com.jajaja.domain.team.dto.response.TeamProductListResponseDto;
 import com.jajaja.domain.team.service.TeamCommandService;
 import com.jajaja.domain.team.service.TeamQueryService;
 import com.jajaja.global.apiPayload.ApiResponse;
-import com.jajaja.global.config.security.annotation.Auth;
+import com.jajaja.global.security.annotation.Auth;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
