+++ src/main/java/com/jajaja/global/S3/service/S3Service.java
@@ -43,8 +43,6 @@ public class S3Service {
                 .withMethod(HttpMethod.PUT)
                 .withExpiration(expiration);
 
-        generatePresignedUrlRequest.addRequestParameter("x-amz-acl", "public-read");
-
         String key = generatePresignedUrlRequest.getKey();
 
         return PresignedUrlUploadResponseDto.builder()
@@ -67,8 +65,6 @@ public class S3Service {
                     GeneratePresignedUrlRequest generatePresignedUrlRequest = new GeneratePresignedUrlRequest(bucket, keyName)
                             .withMethod(HttpMethod.PUT)
                             .withExpiration(expiration);
-
-                    generatePresignedUrlRequest.addRequestParameter("x-amz-acl", "public-read");
 
                     String key = generatePresignedUrlRequest.getKey();
