+++ src/main/java/com/jajaja/global/security/SecurityConfig.java
@@ -1,12 +1,12 @@
-package com.jajaja.global.config.security;
+package com.jajaja.global.security;
 
 import com.jajaja.domain.auth.service.CustomOAuth2UserService;
 import com.jajaja.global.config.CorsConfig;
-import com.jajaja.global.config.security.jwt.JwtAuthenticationEntryPoint;
-import com.jajaja.global.config.security.jwt.JwtAuthenticationFilter;
-import com.jajaja.global.config.security.jwt.JwtExceptionFilter;
-import com.jajaja.global.config.security.jwt.JwtProvider;
-import com.jajaja.global.config.security.oauth.CustomOAuth2SuccessHandler;
+import com.jajaja.global.security.jwt.JwtAuthenticationEntryPoint;
+import com.jajaja.global.security.jwt.JwtAuthenticationFilter;
+import com.jajaja.global.security.jwt.JwtExceptionFilter;
+import com.jajaja.global.security.jwt.JwtProvider;
+import com.jajaja.global.security.oauth.CustomOAuth2SuccessHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
