+++ src/main/java/com/jajaja/global/security/resolver/AuthArgumentResolver.java
@@ -1,7 +1,7 @@
-package com.jajaja.global.config.security.resolver;
+package com.jajaja.global.security.resolver;
 
 import com.jajaja.domain.auth.dto.CustomOAuth2User;
-import com.jajaja.global.config.security.annotation.Auth;
+import com.jajaja.global.security.annotation.Auth;
 import org.springframework.core.MethodParameter;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
